refactor: Reworked objects to pass parcelable objects instead of just IDs

+++ app/src/main/java/com/automotive/hhi/mileagetracker/model/data/Car.java
@@ -18,7 +18,7 @@ public class Car implements Parcelable {
     public Car(){}
 
     public Car(Parcel in){
-        setId(in.readInt());
+        setId(in.readLong());
         setName(in.readString());
         setMake(in.readString());
         setModel(in.readString());
